Remove cirq deploy key (#212)
Now that cirq and openfermion-cirq are public, we don't need this hacky method for allowing travis-ci to access cirq when building openfermioncirq. Also, it's setting off github's warning systems to have a private RSA key in a public repo (as it should! I was surprised that it didn't do it when private to be honest.).

I bet we keep hearing about this for years, as people run scans over historical commits. I look forward to being a false positive from here on out.